requirements-test.txt
-r requirements.txt
coverage
autopep8
ipdb
isort
pep8
pycodestyle
pylint
pylint-django
pylint-plugin-utils
safety